server/requirements.txt
absl-py==0.8.1
appdirs==1.4.4
asgiref==3.4.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.4
Babel==2.8.0
black==20.8b1
boto3==1.14.63
botocore==1.17.63
Brlapi==0.6.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
constantly==15.1.0
cryptography==2.6.1
cssselect==1.1.0
cupshelpers==1.0
distro-info==0.21
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-storages==1.10.1
djangorestframework==3.11.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
fastapi==0.66.0
filelock==3.0.12
Flask==1.1.2
future==0.17.1
gast==0.2.2
gcloud==0.17.0
google-pasta==0.1.7
googleapis-common-protos==1.51.0
grpcio==1.24.1
gunicorn==20.0.4
gym==0.14.0
h11==0.12.0
h5py==2.10.0
httplib2==0.11.3
huggingface-hub==0.0.12
hyperlink==19.0.0
idna==2.10
importlib-metadata==4.6.1
incremental==17.5.0
isort==4.3.21
itemadapter==0.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
jws==0.1.3
kazam==1.4.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.4.3
louis==3.8.0
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
neat-python==0.92
nltk==3.6.2
numpy==1.17.0
oauth2client==3.0.0
opt-einsum==3.1.0
packaging==21.0
pandas==1.1.1
parsel==1.6.0
pathspec==0.8.0
pipenv==11.9.0
Protego==0.1.16
protobuf==3.10.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycryptodome==3.4.3
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.8.2
PyDispatcher==2.0.5
pygame==1.9.6
pyglet==1.3.2
PyGObject==3.30.4
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
Pyrebase==3.0.27
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
python-jwt==2.0.1
python-xlib==0.23
pytz==2020.1
pyxdg==0.25
PyYAML==5.4.1
queuelib==1.5.0
regex==2020.7.14
reportbug===7.5.3-deb10u1
requests==2.24.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.45
safeeyes==2.0.9
scipy==1.3.1
Scrapy==2.2.1
SecretStorage==2.3.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
sshuttle==0.78.5
starlette==0.14.2
tensorflow-estimator==2.0.0
termcolor==1.1.0
tokenizers==0.10.3
toml==0.10.1
tqdm==4.61.2
transformers==4.8.2
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
unattended-upgrades==0.1
urllib3==1.25.10
uvicorn==0.14.0
virtualenv==16.6.2
virtualenv-clone==0.3.0
w3lib==1.22.0
webvtt-py==0.4.6
Werkzeug==0.16.0
wrapt==1.11.2
youtube-dl==2021.6.6
zipp==3.5.0
zope.interface==5.1.0
